dom/tests/mochitest/geolocation/test_errorcheck.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=684722
-->
<head>
  <title>Test for ErrorChecking </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=684722">Mozilla Bug 684722</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

resume_geolocationProvider(function() {
  force_prompt(true, test1);
});

function test1() {
    send404_geolocationProvider(test2);
}

function errorCallback(error) {
    // PositionError has no interface object, so we can't get constants off that.
    is(error.code, error.POSITION_UNAVAILABLE,
       "Geolocation error handler fired");
    is(error.POSITION_UNAVAILABLE, 2,
       "Value of POSITION_UNAVAILABLE should be correct");
    SimpleTest.finish();
}

function successCallback(position) {
    test2();
}

function test2() {
  navigator.geolocation.getCurrentPosition(successCallback, errorCallback);
}
</script>
</pre>
</body>
</html>
